aider: Modified the `Workshop` component to display the name, current number of participants, and maximum number of participants.
# Aider chat conversation:

USER: Great, the workshop components appear in the <main> sector but the contents are empty. For each component/result we want to get the values inside the "properties" object, currently called "Name" (for the name of the workshop), "Current" (for the amount of participants currently in the workshop) and "Max" (for the amount of maximum participants that should be able to attend that workshop).

ASSISTANT: Sure, we can modify the `Workshop` component to display these properties.

Here is the *edit block* for the changes in the `components/Workshop.js` file:

```javascript
components/Workshop.js
<<<<<<< HEAD
import React from 'react';

const Workshop = ({ workshop }) => {
  return (
    <div>
      <h2>{workshop.title}</h2>
      <p>{workshop.description}</p>
    </div>
  );
};

export default Workshop;
=======
import React from 'react';

const Workshop = ({ workshop }) => {
  const { Name, Current, Max } = workshop.properties;
  return (
    <div>
      <h2>{Name.title[0].text.content}</h2>
      <p>Current Participants: {Current.number}</p>
      <p>Max Participants: {Max.number}</p>
    </div>
  );
};

export default Workshop;
>>>>>>> updated
```

This `Workshop` component now receives a `workshop` prop and renders the name of the workshop, the current number of participants, and the maximum number of participants. Please adjust the fields according to the actual structure of your `workshop` object if needed.