aider: Change the `getServerSideProps` function to `getStaticProps` and add a `revalidate` property to the returned object, setting it to 60 seconds.
# Aider chat conversation:

USER: I want to periodically check for changes in the content fetched from Notion and refresh the results if there are new changes.

ASSISTANT: To periodically check for changes in the content fetched from Notion and refresh the results if there are new changes, we can use a combination of `useEffect` and `useState` hooks in React.

We will need to modify the `pages/index.js` file.

Here are the changes:

^C KeyboardInterrupt

USER: I'm using nextJS

ASSISTANT: In a Next.js application, you can use Incremental Static Regeneration (ISR) to periodically update the page data. ISR allows you to update existing pages by re-rendering them in the background as traffic comes in.

To implement this, we need to modify the `pages/index.js` file. Specifically, we need to change the `getServerSideProps` function to `getStaticProps` and add a `revalidate` property to the returned object. The `revalidate` property should be set to the number of seconds you want Next.js to wait before re-generating the page.

Here's the change:

```javascript
pages/index.js
<<<<<<< HEAD
export async function getServerSideProps() {
  const databaseContents = await notion.databases.query({
    database_id: databaseId,
  });
  return {
    props: {
      databaseContents,
    },
  };
}
=======
export async function getStaticProps() {
  const databaseContents = await notion.databases.query({
    database_id: databaseId,
  });
  return {
    props: {
      databaseContents,
    },
    revalidate: 60, // Re-generate the page every 60 seconds
  };
}
>>>>>>> updated
```

This change will make Next.js re-generate the page every 60 seconds. You can adjust the `revalidate` value to suit your needs.