aider: Set the width and height of the segments to 100% to make the squares occupy the whole container.
# Aider chat conversation:

USER: I want those five squares to not have a set width or height but just occupy the whole container

ASSISTANT: To make the squares occupy the whole container, we need to modify the CSS module associated with the `ProgressBar` component. We will set the width and height of the segments to 100% so that they take up the full space of their parent container.

Here is the change:

```css
components/ProgressBar.module.css
<<<<<<< HEAD
.progressBar {
  display: flex;
  justify-content: space-between;
  width: 100%;
}

.segment {
  width: 20px;
  height: 20px;
  background-color: #f3f3f3;
}

.active {
  background-color: #4caf50;
}
=======
.progressBar {
  display: flex;
  justify-content: space-between;
  width: 100%;
  height: 100%;
}

.segment {
  flex: 1;
  background-color: #f3f3f3;
}

.active {
  background-color: #4caf50;
}
>>>>>>> updated
```

This change will make each segment of the progress bar (each square) take up an equal portion of the progress bar's total width and height.