aider: Revert the height of the squares to a normal value and not extend full-width.
# Aider chat conversation:

USER: I made a mistake, I want those squares to have a normal height, not extend full-width.

ASSISTANT: To revert the height of the squares to a normal value and not extend full-width, we need to modify the CSS module associated with the `ProgressBar` component. We will set the height of the segments to a fixed value.

Here is the change:

```css
components/ProgressBar.module.css
<<<<<<< HEAD
.progressBar {
  display: flex;
  justify-content: space-between;
  width: 100%;
  height: 100%;
}

.segment {
  flex: 1;
  background-color: #f3f3f3;
}

.active {
  background-color: #4caf50;
}
=======
.progressBar {
  display: flex;
  justify-content: space-between;
  width: 100%;
}

.segment {
  flex: 1;
  height: 20px;
  background-color: #f3f3f3;
}

.active {
  background-color: #4caf50;
}
>>>>>>> updated
```

This change will make each segment of the progress bar (each square) take up an equal portion of the progress bar's total width and have a fixed height of 20px.