aider: Add padding between the squares of the progress bar.
# Aider chat conversation:

USER: I want those squares to have a bit of padding in between

ASSISTANT: To add padding between the squares, we need to modify the CSS module associated with the `ProgressBar` component. We will add a `margin` property to the `.segment` class.

Here is the change:

```css
components/ProgressBar.module.css
<<<<<<< HEAD
.progressBar {
  display: flex;
  justify-content: space-between;
  width: 100%;
}

.segment {
  flex: 1;
  height: 20px;
  background-color: #f3f3f3;
}

.active {
  background-color: #4caf50;
}
=======
.progressBar {
  display: flex;
  justify-content: space-between;
  width: 100%;
}

.segment {
  flex: 1;
  height: 20px;
  background-color: #f3f3f3;
  margin: 0 2px;
}

.active {
  background-color: #4caf50;
}
>>>>>>> updated
```

This change will add a 2px margin to the left and right of each segment of the progress bar (each square), creating a bit of padding between them.