aider: Set the width and height of the progress bar segments to be equal, creating square divs.
# Aider chat conversation:

USER: I want each div inside the progressbar to be a square, always with the height matching the current width.

ASSISTANT: To achieve this, we need to modify the CSS for the progress bar segments. We will set the width and height to be equal and use a percentage value to ensure it's responsive.

We will modify the `components/ProgressBar.module.css` file.

Here is the edit block for the changes:

```css
components/ProgressBar.module.css
<<<<<<< HEAD
.segment {
  flex: 1;
  height: 20px;
  background-color: #f3f3f3;
  margin: 0 2px;
}
=======
.segment {
  flex: 1;
  width: 20%;
  height: 0;
  padding-bottom: 20%;
  background-color: #f3f3f3;
  margin: 0 2px;
}
>>>>>>> updated
```

In the updated code, we set the width and height to be equal (20% of the parent's width). We use `padding-bottom` instead of `height` to create a square that maintains its aspect ratio. The `padding-bottom` is relative to the width of the element, so it will always be square regardless of the width.