aider: Set the font size of the title to be 3% of the viewport's width.
# Aider chat conversation:

USER: I want the size of the title to be bigger but dynamic so it adapts to the given size (roughly one third of the container)

ASSISTANT: To achieve this, we can use viewport units in CSS. The `vw` unit represents a percentage of the viewport's width, so it will scale with the size of the window.

We will modify the `components/Workshop.module.css` file.

Here is the edit block for the changes:

```css
components/Workshop.module.css
<<<<<<< HEAD
.leftColumn {
  flex: 1;
  display: flex;
  flex-direction: column;
  justify-content: center;
  text-transform: uppercase;
  font-weight: lighter;
}
=======
.leftColumn {
  flex: 1;
  display: flex;
  flex-direction: column;
  justify-content: center;
  text-transform: uppercase;
  font-weight: lighter;
}

.leftColumn h2 {
  font-size: 3vw;
}
>>>>>>> updated
```

In the updated code, we set the font size of the title (`h2`) to be 3% of the viewport's width. This will make the title size dynamic and roughly one third of the container's width.