aider: Increased the padding of the workshop module and the spacing in between the segments of the progress bar.
# Aider chat conversation:

USER: I want to increase the padding of the workshop module as well as the spacing in between the segments of the progressbar

ASSISTANT: To increase the padding of the workshop module, we need to modify the `components/Workshop.module.css` file.

To increase the spacing between the segments of the progress bar, we need to modify the `components/ProgressBar.module.css` file.

Here are the edit blocks for the changes:

```css
components/Workshop.module.css
<<<<<<< HEAD
.workshop {
  display: flex;
  justify-content: space-between;
  background-color: black;
  color: white;
  margin: 10px;
  padding: 10px;
  flex: 1;
  width: 100%;
}
=======
.workshop {
  display: flex;
  justify-content: space-between;
  background-color: black;
  color: white;
  margin: 20px;
  padding: 20px;
  flex: 1;
  width: 100%;
}
>>>>>>> updated
```

In the updated code, we increased the margin and padding of the workshop module from 10px to 20px.

```css
components/ProgressBar.module.css
<<<<<<< HEAD
.segment {
  flex: 1;
  width:
^C KeyboardInterrupt

USER: I need those margins to also be in dynamic values

ASSISTANT: To make the margins and padding dynamic, we can use viewport units in CSS. The `vw` unit represents a percentage of the viewport's width, so it will scale with the size of the window.

Here are the edit blocks for the changes:

```css
components/Workshop.module.css
<<<<<<< HEAD
.workshop {
  display: flex;
  justify-content: space-between;
  background-color: black;
  color: white;
  margin: 20px;
  padding: 20px;
  flex: 1;
  width: 100%;
}
=======
.workshop {
  display: flex;
  justify-content: space-between;
  background-color: black;
  color: white;
  margin: 2vw;
  padding: 2vw;
  flex: 1;
  width: 100%;
}
>>>>>>> updated
```

In the updated code, we set the margin and padding of the workshop module to be 2% of the viewport's width.

```css
components/ProgressBar.module.css
<<<<<<< HEAD
.segment {
  flex: 1;
  width: 20%;
  height: 0;
  padding-bottom: 20%;
  background-color: #f3f3f3;
  margin: 0 2px;
}
=======
.segment {
  flex: 1;
  width: 20%;
  height: 0;
  padding-bottom: 20%;
  background-color: #f3f3f3;
  margin: 0 0.2vw;
}
>>>>>>> updated
```

In the updated code, we set the margin of the progress bar segments to be 0.2% of the viewport's width.

USER: InvalidEditBlock: edit failed!

components/Workshop.module.css does not contain the *exact sequence* of HEAD lines you specified.
Try again.
DO NOT skip blank lines, comments, docstrings, etc!
The HEAD block needs to be EXACTLY the same as the lines in components/Workshop.module.css with nothing missing!

components/Workshop.module.css does not contain these 10 exact lines in a row:
```
.workshop {
  display: flex;
  justify-content: space-between;
  background-color: black;
  color: white;
  margin: 20px;
  padding: 20px;
  flex: 1;
  width: 100%;
}
```

ASSISTANT: I apologize for the confusion. Let's correct the edit block for the `components/Workshop.module.css` file. The current margin and padding values are 10px, not 20px. Here's the corrected edit block:

```css
components/Workshop.module.css
<<<<<<< HEAD
.workshop {
  display: flex;
  justify-content: space-between;
  background-color: black;
  color: white;
  margin: 10px;
  padding: 10px;
  flex: 1;
  width: 100%;
}
=======
.workshop {
  display: flex;
  justify-content: space-between;
  background-color: black;
  color: white;
  margin: 2vw;
  padding: 2vw;
  flex: 1;
  width: 100%;
}
>>>>>>> updated
```

And the edit block for the `components/ProgressBar.module.css` file is correct as provided earlier.