aider: Changed workshop columns to stack vertically for tablet-portrait and mobile viewports.
# Aider chat conversation:

USER: For tablet-portrait and mobile viewports I would like the workshop columns to stack vertically, instead of being left and right.

ASSISTANT: To achieve this, we need to modify the CSS for the `Workshop` component. We will use media queries to change the flex direction to column for smaller viewports.

Here is the edit block for the `components/Workshop.module.css` file:

```css
components/Workshop.module.css
<<<<<<< HEAD
.workshop {
  display: flex;
  justify-content: space-between;
  background-color: black;
  color: white;
  margin: 0.5%;
  padding: 1% 3%;
  flex: 1;
  width: 100%;
  border-radius: 2vw;
}
=======
.workshop {
  display: flex;
  justify-content: space-between;
  background-color: black;
  color: white;
  margin: 0.5%;
  padding: 1% 3%;
  flex: 1;
  width: 100%;
  border-radius: 2vw;
}

@media (max-width: 768px) {
  .workshop {
    flex-direction: column;
  }
}
>>>>>>> updated
```

This change will make the workshop columns stack vertically on viewports that are 768px wide or smaller, which typically includes tablet-portrait and mobile viewports.